+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,13 @@
 # stories
 
+## 0.8.16
+
+### Patch Changes
+
+- Adds form column for web
+- Updated dependencies
+  - @shallot-ui/web@0.41.16
+
 ## 0.8.15
 
 ### Patch Changes